ci: don't run tests for "openssl buggy" ruby 2.3
There are a couple of important OpenSSL::PKey::EC bugs in ruby 2.3,
which were fixed in ruby 2.4+ in 2016 but were never backported to ruby
2.3.

As of today (2018 May) ruby 2.3 is on "security maintenance phase",
which means it's no longer accepting bug fixes.

Given the above, and that this gem is using OpenSSL::PKey::EC at it's
core, it's probably ok to drop explicit support for ruby 2.3.

Refs:

* [Let OpenSSL::PKey::EC follow the general PKey interface](https://bugs.ruby-lang.org/issues/6567)
* [OpenSSL::X509::Request can't sign() an OpenSSL::PKey::EC](https://bugs.ruby-lang.org/issues/5600)
* [Ruby 2.3.7 Released](https://www.ruby-lang.org/en/news/2018/03/28/ruby-2-3-7-released/)