feat: support 'android-safetynet' attestation format
Verify Android SafetyNet attestation as described at:
- https://www.w3.org/TR/2018/CR-webauthn-20180807/#android-safetynet-attestation
- https://developer.android.com/training/safetynet/attestation#compat-check-response

This is the attestation returned by Chrome 70 on Android when using the fingerprint scanner to register.

The W3C documentation does a poor job of explaining how and why the certificates must be verified. If this step were to be skipped, anybody could forge a response with a self-signed certificate issued to attest.android.com. From what I could find only Google's checklist (https://developer.android.com/training/safetynet/attestation-checklist) as well as some third-party resources mention it.

Also it is not really clear what a "valid SafetyNet response of version ver" is supposed to look like.