fix: verify U2F attested credential data AAGUID is zeroed out
The conformance test tool notified me of this. It used to be called out more clearly in the spec:
https://github.com/w3c/webauthn/pull/539/commits/4a376bf198f839abe123e5e32458c83d40b67f29
but that has been changed over time.

Now https://www.w3.org/TR/2019/PR-webauthn-20190117/#fido-u2f-attestation mentions only:
> Extract the claimed rpIdHash from authenticatorData, and the claimed
> credentialId and credentialPublicKey from authenticatorData.attestedCredentialData.
which is vague about what you're supposed to expect.

Fortunately we have
https://fidoalliance.org/specs/fido-v2.0-id-20180227/fido-client-to-authenticator-protocol-v2.0-id-20180227.html#u2f-authenticatorMakeCredential-interoperability
confirming the AAGUID should be initialized with 16 zeroes.

The actual AAGUID for U2F authenticators lives in the certificates, similar to the packed attestation format.