refactor: make required credentials creation options more explicit
These fields are required by the spec, they have meaning in how and when they
should be used: https://www.w3.org/TR/2019/PR-webauthn-20190117/#dictionary-pkcredentialentity

I think current the defaults are not helpful for users of the gem since they may
be shown by the browser in the future. If this happens, the unchanged defaults
may lead to authentication prompts like:
> Scan fingerprint to authenticate as web-user for web-server

To not break backwards compatibility while making the need for explicit
configuration more obvious, I've made the keyword arguments optional for now
with a note to change this for version 2.0.