refactor: changed allowed_credential to just take one public_key
This is a breaking change.

The reason to push the responsibility up to the application to find the right public_key for a given credential ID is that applications that want to do signature counter verification and updating (a "SHOULD" in the standard) need to do so anyway. It doesn't make much sense to have both the gem and the app find the right credential independently.

The AuthenticatorAssertionResponse object doesn't need to know about the credential ID anymore either.