feat: pass safetynet attestation remaining conformance test
This is a temporary solution to pass P-1 android-safetynet attestation
conformance test. Right now the certificate used by the test cannot be
verified
(https://github.com/fido-alliance/conformance-tools-issues/issues/518),
which results in the other validations not being tested.
The trust store used to verify the certificate chain should be updated
not to use the `default_paths` once the problem with the conformance
tool is fixed.