test: patch in TPM conformance test behaviour
There's only one valid manufacturing ID in conformance testing.

Intercept the call with the TCG trust store object, replace that with the one from the metadata statement, and continue with the verification code as-is.

All of this is never needed in real life so let's not pollute our production code with it.