feat: raise an error if `legacy_u2f_appid` is not set
If the appid extension was requested and the output was `true` but the
`legacy_u2f_appid` config was not set, raise an error.

We were previously were passing the `rp_id` instead if the mentioned
config was not present, but that is needlessly given that it will fail
with a rp verification error.