fix: do not add the key to the hash if its value is `nil`
In Rails, `nil` responds to `as_json` therefore the attributes whose
value is `nil` are added to the result of `as_json`, which may end up
causing issues with some authenticators – see https://github.com/cedarcode/webauthn-ruby/issues/441.

This commits adds another condition to the if in order to prevent
this from happening.