Update conformance specs to use FIDO MDS v3 (#407)
* build(conformanceTests): bump `webauthn` version

* build(conformanceTests): bump `byebug`

* build(conformanceTests): bump `fido_metadata` version

* build(conformanceTests): use ruby `3.4.2` for conformance specs

* build(conformanceTests): update bundler

* build(conformanceTests): add `webrick`

We have to add `webrick` too as it was removed from Ruby's standard
library in 3.0 (https://bugs.ruby-lang.org/issues/17303).

* feature(conformanceTests): use MDS v3

* build(conformanceTests): point `fido_metadata` to its repo's `main` branch

* build(conformanceTests): bump `sinatra` from `2.2.4` to `4.1.1`

* build(conformanceTests): add `puma` and `rackup`

Fixes an error where `sinatra` was not being able to start.

```
Sinatra could not start, the required gems weren't found!

Add them to your bundle with:

    bundle add rackup puma

or install them with:

    gem install rackup puma
```